add support for web to batch upload multicam datasets
